requirements.txt

playwright
python-dotenv
pandas
